Add support for eslint & stylelint, cleanup css files to be pure css
