+++ .eslintrc
@@ -1,0 +1,38 @@
+{
+    "env": {
+      "browser": true,
+      "es2021": true
+    },
+    "extends": [
+      "@stackla/eslint-config",
+      "eslint:recommended",
+      "plugin:promise/recommended",
+      "plugin:@typescript-eslint/recommended",
+      "plugin:eslint-plugin-css/recommended"
+    ],
+    "parser": "@typescript-eslint/parser",
+    "parserOptions": {
+      "ecmaVersion": 12,
+      "sourceType": "module"
+    },
+    "plugins": [
+      "@typescript-eslint",
+      "promise"
+    ],
+    "rules": {
+      "promise/prefer-await-to-then": "error",
+      "@typescript-eslint/no-explicit-any": "error",
+      "@typescript-eslint/consistent-type-assertions": ["warn", { "assertionStyle": "never" }]
+    },
+    "overrides": [
+      {
+        "files": ["./src/**/*.ts"],
+        "rules": {
+          "@typescript-eslint/no-floating-promises": "error"
+        },
+        "parserOptions": {
+          "project": true
+        }
+      }
+    ]
+  }
