+++ .npmrc
@@ -1,0 +1,2 @@
+@stackla:registry=https://npm.pkg.github.com
+//npm.pkg.github.com/:_authToken=${NODE_AUTH_TOKEN}
