+++ .stylelintrc.json
@@ -1,0 +1,11 @@
+{
+    "extends": [
+      "stylelint-config-standard",
+      "stylelint-prettier/recommended"
+    ],
+    "rules": {
+      "prettier/prettier": true,
+      "declaration-no-important": true,
+      "no-empty-source": null
+    }
+  }
