+++ jest.config.js
@@ -1,0 +1,3 @@
+const baseConfig = require("@stackla/jest-config");
+
+module.exports = baseConfig({ tsConfig: require("./tsconfig.json") });
