+++ package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@stackla/widget-templates",
+  "version": "0.0.1",
+  "description": "Widget Templates",
+  "scripts": {
+    "lint": "stylelint '**/*.css'",
+    "test": "jest --color --passWithNoTests",
+    "build": "tsc -p tsconfig.build.json"
+  },
+  "license": "MIT",
+  "devDependencies": {
+    "@stackla/eslint-config": "^1.0.0",
+    "@stackla/jest-config": "^1.0.6",
+    "@stackla/types": "^0.0.6",
+    "@typescript-eslint/eslint-plugin": "^7.13.1",
+    "eslint": "^8.56.0",
+    "eslint-config-stylelint": "^21.0.0",
+    "eslint-plugin-css": "^0.9.2",
+    "eslint-plugin-promise": "^6.2.0",
+    "stylelint": "^16.6.1",
+    "stylelint-config-standard": "^36.0.1",
+    "stylelint-prettier": "^5.0.0",
+    "typescript": "^5.5.0"
+  }
+}
