+++ src/widgets/example/components/add-to-cart/base.css
@@ -1,0 +1,4 @@
+.ugc-add-to-cart-button {
+  max-width: 300px;
+  text-transform: uppercase;
+}
