+++ src/widgets/example/components/add-to-cart/base.template.ts
@@ -1,5 +1,5 @@
 export default () => {
-    return `
+  return `
     <div id="stacklapopup-add-to-cart"></div>
   `;
 };
