+++ src/widgets/example/components/cross-sellers/base.template.ts
@@ -1,3 +1,3 @@
 export default () => {
-    return ``;
+  return ``;
 };
