+++ src/widgets/example/components/expanded-tile/base.template.ts
@@ -1,20 +1,20 @@
-import Sdk from "../../../../types/sdk";
-import {Tile} from "../../../../types/tile";
+import type { Sdk, Tile } from '@stackla/types';
 
 export default (sdk: Sdk) => {
-    const tile = sdk.tiles.getTile() as Tile;
-    const shopspotEnabled = sdk.isComponentLoaded('shopspots');
-    const productsEnabled = sdk.isComponentLoaded('products');
-    const parent = sdk.getNodeId();
-    return `<div class="panel">
+  const tile = sdk.tiles.getTile() as Tile;
+  const shopspotEnabled = sdk.isComponentLoaded("shopspots");
+  const productsEnabled = sdk.isComponentLoaded("products");
+  const parent = sdk.getNodeId();
+
+  return `<div class="panel">
         <a class="exit" href="#">Close</a>
         <div class="panel-left">
             <div class="image-wrapper">
                 <div class="image-wrapper-inner">
                     <div class="image">
-                        ${shopspotEnabled ? `<shopspot-flyout parent="${parent}"></shopspot-flyout>` : ''}
-                        ${shopspotEnabled ? `<shopspot-icon parent="${parent}"/></shopspot-icon>` : ''}
-                        ${tile.image ? `<img class="image-element" src="${tile.image}" />` : ''}
+                        ${shopspotEnabled ? `<shopspot-flyout parent="${parent}"></shopspot-flyout>` : ""}
+                        ${shopspotEnabled ? `<shopspot-icon parent="${parent}"/></shopspot-icon>` : ""}
+                        ${tile.image ? `<img class="image-element" src="${tile.image}" />` : ""}
                     </div>
                 </div>
             </div>
@@ -31,17 +31,17 @@ export default (sdk: Sdk) => {
                         <div class="user-info-wrapper">
                             <div class="user-info">
                                 ${
-        tile.avatar
-            ? `<span class="avatar-wrapper">
+                                  tile.avatar
+                                    ? `<span class="avatar-wrapper">
                                     <a class="avatar-link" href="${tile.original_url}" target="_blank">
                                        <img src="${tile.avatar}">
                                     </a>
                                 </span>`
-            : ''
-    }
+                                    : ""
+                                }
                                 ${
-        tile.user
-            ? `<a class="user-link" href="${tile.original_url}" target="_blank">
+                                  tile.user
+                                    ? `<a class="user-link" href="${tile.original_url}" target="_blank">
                                     <div class="user-top">
                                         <span class="user-name">${tile.user}</span>
                                     </div>
@@ -49,13 +49,13 @@ export default (sdk: Sdk) => {
                                         <span class="user-handle">@${tile.user}</span>
                                     </div>
                                 </a>`
-            : ''
-    }
+                                    : ""
+                                }
                             </div>
                         </div>
                         <div class="caption">
-                            <p class="caption-paragraph">${tile.message ?? ''}</p>
-                            ${productsEnabled ? `<ugc-products parent="${parent}">` : ''}
+                            <p class="caption-paragraph">${tile.message ?? ""}</p>
+                            ${productsEnabled ? `<ugc-products parent="${parent}">` : ""}
                         </div>
                         <div class="footer">
                             <span class="base-v2 source source-instagram">
