+++ src/widgets/example/layout.hbs
