+++ src/widgets/example/tile.hbs
