--- tsconfig.json
@@ -1,40 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": ".",
-    "target": "es6",
-    "module": "commonjs",
-    "lib": [
-      "es2019",
-      "es2021",
-      "DOM",
-      "DOM.Iterable",
-      "dom"
-    ],
-    "declaration": true,
-    "declarationMap": true,
-    "sourceMap": false,
-    "inlineSources": false,
-    "outDir": "./",
-    "strict": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noImplicitReturns": true,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitAny": true,
-    "noImplicitThis": true,
-    "moduleResolution": "node",
-    "allowSyntheticDefaultImports": true,
-    "experimentalDecorators": true,
-    "forceConsistentCasingInFileNames": true,
-    "noImplicitOverride": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "resolveJsonModule": true,
-    "paths": {
-      "@libs/*": ["./src/libs/*"],
-      "@tests/*": ["./tests/*"]
-    }
-  },
-  "include": ["src/**/*.ts"],
-  "exclude": []
-}
