--- widgets/example/components/add-to-cart/base.styles.ts
@@ -1,6 +1,0 @@
-export default `
-  .ugc-add-to-cart-button {
-    max-width: 300px !important;
-    text-transform: uppercase;
-  }
-`;
