--- widgets/example/widget.css
@@ -1,32 +1,0 @@
-.hd {
-    font-size:16px;
-}
-.ugc-tile,.tile {
-    display: inline-block;
-    vertical-align: top;
-    width: 19%;
-    margin: 0 auto;
-    text-align: center;
-}
-.ugc-tile img {
-    height: 250px;
-    width: 250px;
-}
-.ugc-tile p {
-    min-height: 15px;
-    max-height: 200px;
-    overflow: hidden;
-}
-.loader {
-    padding: 20px;
-    margin: 0 auto;
-    width: 50px;
-    aspect-ratio: 1;
-    border-radius: 50%;
-    background: radial-gradient(farthest-side, #b4b2af 94%, #0000) top / 8px 8px no-repeat, conic-gradient(#0000 30%, #928f8a);
-    -webkit-mask: radial-gradient(farthest-side, #0000 calc(100% - 8px), #000 0);
-    animation: l13 1s infinite linear;
-}
-@keyframes l13{
-    100%{transform: rotate(1turn)}
-}
